grinn-auth-service/src/main/java/ru/grinncorp/auth/service/security/OraUserDetailsService.java
package ru.grinncorp.auth.service.security;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import ru.grinncorp.auth.repository.UserRepository;

@Service
public class OraUserDetailsService implements UserDetailsService {
	
	@Autowired
	private UserRepository repository;

	@Override
	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
		
		return repository.findById(username).orElseThrow(() -> new UsernameNotFoundException(username));
	}
	

}
